+++ decoders/deye_can/__init__.py
@@ -1,0 +1,5 @@
+'''
+Deye CAN protocol decoder.
+'''
+
+from .pd import Decoder